Fikset liten bokstav
