+++ src/main/java/inf112/balmerspeak/app/robot/Robot.java
@@ -2,6 +2,7 @@ package inf112.balmerspeak.app.robot;
 
 import inf112.balmerspeak.app.cards.Deck;
 import inf112.balmerspeak.app.cards.ProgramCard;
+import inf112.balmerspeak.app.cards.Rotation;
 
 import java.util.ArrayList;
 
@@ -83,6 +84,29 @@ public class Robot implements IRobot{
         for (ProgramCard cards : deck.getCards(amount))
             hand.add(cards);
         return hand;
+    }
+
+    public Direction turn(Rotation rotation, Direction direction) {
+        switch (direction) {
+            case NORTH:
+                if (rotation.equals(Rotation.right)) return Direction.EAST;
+                else if (rotation.equals(Rotation.uturn)) return Direction.SOUTH;
+                else return Direction.WEST;
+            case SOUTH:
+                if (rotation.equals(Rotation.right)) return Direction.WEST;
+                else if (rotation.equals(Rotation.uturn)) return Direction.NORTH;
+                else return Direction.EAST;
+            case WEST:
+                if (rotation.equals(Rotation.right)) return Direction.NORTH;
+                else if (rotation.equals(Rotation.uturn)) return Direction.EAST;
+                else return Direction.SOUTH;
+            case EAST:
+                if (rotation.equals(Rotation.right)) return Direction.SOUTH;
+                else if (rotation.equals(Rotation.uturn)) return Direction.WEST;
+                else return Direction.NORTH;
+            default:
+                return null;
+        }
     }
